feat(#450): display slide name for the navigation in the remote
