+++ studio/src/app/modals/editor/app-slide-navigate/app-slide-navigate.tsx
@@ -1,6 +1,8 @@
 import {Component, Listen, Element, State, h, EventEmitter, Event} from '@stencil/core';
 import {ItemReorderEventDetail} from '@ionic/core';
 
+import {findSlidesTitle} from '@deckdeckgo/deck-utils';
+
 @Component({
     tag: 'app-slide-navigate',
     styleUrl: 'app-slide-navigate.scss'
@@ -10,14 +12,14 @@ export class AppSlideNavigate {
     @Element() el: HTMLElement;
 
     @State()
-    slides: string[];
+    private slides: string[];
 
     @Event() private reorder: EventEmitter<ItemReorderEventDetail>;
 
     async componentDidLoad() {
         history.pushState({modal: true}, null);
 
-        this.slides = await this.getSlidesTitle();
+        this.slides = await findSlidesTitle();
     }
 
     @Listen('popstate', { target: 'window' })
@@ -31,47 +33,6 @@ export class AppSlideNavigate {
 
     async jumpToSlide(index: number) {
         await (this.el.closest('ion-modal') as HTMLIonModalElement).dismiss(index);
-    }
-
-    private getSlidesTitle(): Promise<string[]> {
-        return new Promise<string[]>((resolve) => {
-            if (!document) {
-                resolve();
-                return;
-            }
-
-            const results: string[] = [];
-
-            const slides: NodeListOf<HTMLElement> = document.querySelectorAll('deckgo-deck > *');
-
-            if (slides) {
-                for (const slide of Array.from(slides)) {
-                    if (slide.tagName && slide.tagName.toLowerCase().indexOf('deckgo-slide') > -1) {
-                        const title: HTMLElement = slide.querySelector('[slot="title"]');
-
-                        if (title && title.textContent !== '') {
-                            results.push(title.textContent);
-                        } else {
-                            const start: HTMLElement = slide.querySelector('[slot="start"],[slot="header"]');
-
-                            if (start && start.textContent !== '') {
-                                results.push(start.textContent);
-                            } else {
-                                const end: HTMLElement = slide.querySelector('[slot="end"],[slot="footer"]');
-
-                                if (end && end.textContent !== '') {
-                                    results.push(end.textContent);
-                                } else {
-                                    results.push('');
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            resolve(results);
-        });
     }
 
     private onReorder($event: CustomEvent<ItemReorderEventDetail>): Promise<void> {
