+++ utils/deck/src/index.ts
@@ -1,2 +1,3 @@
-export * from './utils/deckdeckgo-parse-style.utils';
-export * from './utils/deckdeckgo-parse-content.utils';
+export * from './utils/parse-style.utils';
+export * from './utils/parse-content.utils';
+export * from './utils/slides.utils';
